Add `BuildConfig` unnamed constructor
- Separate out parsing into a separate library to reduce noise
- Add a default constructor with named args for the fields.
